Version 2.6.0 prerelease

+++ AppData/Actions/bulkDeleteMessages.js
@@ -15,6 +15,7 @@ module.exports = {
     "menuBar1":{"choices": ["Custom*", "Max"], "storeAs":"amount", "extraField":"quantity"},
     "sepbar2":"", "btext2":"Must Be Sent By",
     "menuBar2": {"choices": ["Message Author", "User*", "Anybody"], "storeAs":"user", "extraField":"fromWho"},
+    "invisible":"",
     "variableSettings": {
         "channelVariable": {
             "Variable*": "direct"
@@ -38,7 +39,7 @@ module.exports = {
         var tempVars = JSON.parse(fs.readFileSync('./AppData/Toolkit/tempVars.json', 'utf8'));
 
         let channel;
-        if (values.channelFrom == 'Message Channel*') {
+        if (values.channelFrom == 'Message Channel') {
             channel = message.channel
         } else {
             const channelID = tempVars[uID][varTools.transf(values.channelVariable, uID, tempVars)].id;
@@ -62,13 +63,18 @@ module.exports = {
 
         let amount = 100;
         if (values.amount == 'Custom*') {
-            let amount = parseFloat(varTools.transf(values.quantity, uID, tempVars))
+            amount = parseFloat(varTools.transf(values.quantity, uID, tempVars))
+        }
+        if (amount > 99) {
+            console.log('Bulk Delete >>> Troublesome Amount')
         }
 
-        let checkFor = m => m.createdTimestamp > timestamp && m.content.includes(mustInclude)
 
-        channel.messages.fetch({limit: amount}).then(messages => {
+        let checkFor = m => m.createdTimestamp > timestamp && m.content.toLowerCase().includes(mustInclude.toLowerCase())
+        var deletedMessages = 0;
+        channel.messages.fetch({limit: 100}).then(messages => {
             messages.filter(checkFor).forEach( msg => {
+                if (deletedMessages > amount) return
                 if (mustBeFrom == null) {
                     msg.delete()
                 } else {
@@ -76,6 +82,7 @@ module.exports = {
                         msg.delete()
                     }
                 }
+                deletedMessages++
             })
             })
